Stops setting default value when an initial value exists
When an initial value was given for an attribute, we'd first set it to its default value, and only then to the initial value. This is not necessary, so removing this step will result in less computations. Also, setting the initial value directly will help when adding the logic for configuring an attribute as `initOnly`.