Changes karma tests to use commonjs instead of browserify
This makes the tests run faster, since we're just running babel instead of that and browserify.
Stack traces are also slightly better now that we're using karma-source-map-support. They works fine for the lines related to test files, pointing to the correct line and column numbers, but stops working when the trace reaches a src file. That's due to the coverage preprocessor and will be fixed later on another commit.