Removes direct calls to private functions on the "dom" test
These functions will be called automatically on the browsers that need them, we shouldn't be forcing them to be called on browsers with native implementations. That doesn't improve coverage, since these functions will already be covered when run on the appropriate browsers, and it violates encapsulation by calling protected functions.