Removes unnecessary code from the jsx example for modal nested components
We were showing the whole Modal.js file for the final jsx example in the tutorial for nested components. That's not necessary though, since we only really want to show the new "render" function, so there's no need to repeat code that the developer has seen other times before. It's best to focus on the code we're currently working on.

Also, it was giving the wrong impression that the `close` method wasn't in the file anymore. Showing just the part that changes will be clearer and without unnecessary repetition.