Attaches listeners when decorating existing elements
This stopped working because we now only handle inline listeners inside "handleInterceptedAttributesCall_", which is only called when the attribute value changes for an element.

When decorating the final value for the attribute in the element is already set though, so incremental dom doesn't update it. We need to manually add these events on the first render.