Makes all validators return functions to be called
Some validators needed to be called to return a real validator function, while others didn't. This is not consistent, and since the sugar api that will be built for state will need everything to be called as well it will be even weirder if the original functions behave differently.

Regardless, if the simpler validators are called with arguments, they will run instead of returning a function, so that they can still be called easily as before.