Makes all validators accept null/undefined values - #140
Now, to stop accepting those the "required" config should be used instead (via the `Config` module).

Also, the `shapeOf` validator can receive config objects, for specifying not only the validator but also if it should be required or not. If this information isn't given, any property in the defined shape will be assumed to be optional.