Removes Mobile Safari from list of saucelabs browsers
Saucelabs is having some kind of problem running tests for Mobile Safari, it's as if it fails due to the number of tests we're trying to run there. The tests themselves seem to be working fine if we try to run them on their own, few at a time.

I'm removing Mobile Safari for now, until we figure out how to fix it in the CI.