Fixes gulp tasks for tests
Developers are used to using our gulp tasks for everything in all metal repos, but for this main repo those tasks would fail, and so it was required to use npm scripts instead.

This fixes the tasks for metal.js's use case so that people don't have to change their workflows when they come to this repo.